+++ Exercism/c/largest-series-product/src/largest_series_product.h
@@ -1,0 +1,8 @@
+#ifndef LARGEST_SERIES_PRODUCT_H
+#define LARGEST_SERIES_PRODUCT_H
+
+#include <string.h>
+
+long long largest_series_product(char* series, size_t chain);
+
+#endif